Updated bad color from darkred to yellow
I am red-green colorblind which affects 6% of the male population.  Having one chart that uses both darkred and darkgreen is incredibly difficult for me to read.  I didn't even notice that there were different colors until I saw the source code.  What I inevitably did was change one of the colors to something much brighter.  I would also be okay with changing any of the dark colors to a bright color.  Dark colors are hard to distinguish from each other in general, and when they are both red and green, then it is especially difficult.